+++ .gitignore
@@ -1,0 +1,2 @@
+teste
+compila e roda.txt
