--- Sanrio-Collector-Card-Game/teste.c
@@ -1,7 +1,0 @@
-#include <stdio.h>
-
-int main(){
-    printf("teste albuablublublu");
-
-    return 0;
-}
