+++ cartas.txt
@@ -1,0 +1,15 @@
+Hello Kitty, 4
+My Melody, 3
+Kuromi, 3
+Cinnamoroll, 3
+Pompompurin, 2
+Pochacco, 2
+Chococat, 2
+Badtz-Maru, 2
+Keroppi, 1
+Tuxedosam, 1
+Dear Daniel, 1
+Gudetama, 0
+Little Twin Stars (Kiki & Lala), 0
+Hangyodon, 0
+Aggrestuko, 0
