+++ resources/LICENSE
@@ -1,0 +1,1 @@
+Assets license.
